requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
blobfile==2.0.1
branca==0.4.2
bs4==0.0.1
category-encoders==2.2.2
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
chromedriver-autoinstaller==0.2.2
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
cryptography==39.0.0
dateparser==1.0.0
dnspython==2.3.0
drf-jwt==1.19.2
exceptiongroup==1.1.0
filelock==3.0.12
Fiona==1.8.20
fonttools==4.38.0
frozenlist==1.3.3
h11==0.14.0
htmldate==0.9.0
idna==3.2
importlib-metadata==4.6.3
iniconfig==1.1.1
Jinja2==3.0.1
kiwisolver==1.3.1
lxml==4.6.3
multidict==6.0.4
munch==2.5.0
nltk==3.6.2
outcome==1.2.0
platformdirs==2.6.2
pluggy==0.13.1
psycopg2==2.9.5
pycparser==2.21
pycryptodomex==3.16.0
pydot==1.4.2
PyJWT==2.6.0
pyparsing==2.4.7
pyproj==3.1.0
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
regex==2021.8.28
requests==2.26.0
selenium==3.141.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tenacity==8.0.1
threadpoolctl==2.2.0
tiktoken==0.1.2
toml==0.10.2
tomli==2.0.1
trio-websocket==0.9.2
typing-extensions==3.10.0.0
tzdata==2022.7
tzlocal==3.0
uritools==3.0.2
urlextract==1.3.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.5.0
absl-py==1.0.0
accumulation-tree==0.6.2
asgiref==3.4.0
astunparse==1.6.0
Brotli==1.0.0
cachetools==4.2.0
category-encoders==2.3.0
certifi==2021.10.0
charset-normalizer==2.0.0
click==8.0.0
cvxpy==1.1.10
cycler==0.11.0
dataclasses==0.0
dill==0.3.0
Django==3.2.0
django-extensions==3.2.1
django-registration==3.0
django-sslserver==0.22
djangorestframework==3.13.0
django-cors-headers==3.13.0
djangotoolbox==1.8.0
dpd-components==0.1.0
drf-jwt==1.19.0
ecos==2.0.00
Flask==2.0.0
Flask-Compress==1.10.0
flatbuffers==1.9
fonttools==4.28.0
future==0.18.0
gast==0.4.0
grpcio==1.41.0
h5py==3.5.0
idna==3.0
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.1.0
kiwisolver==1.3.0
libclang==12.0.0
Markdown==3.3.0
MarkupSafe==2.0.0
mysqlclient==2.1.0
nltk==3.6.0
oauthlib==3.1.0
opt-einsum==3.3.0
ortools==9.2.9970
osqp==0.6.2.post5
packaging==21.0
patsy==0.5.0
Pillow==9.0.0
protobuf==3.19.0
pyasn1==0.4.0
pyasn1-modules==0.2.0
PyMySQL==1.0.0
Pympler==1.0.0
pyOpenSSL==22.1.0
pyparsing==3.0.0
python-dateutil==2.8.0
pytz==2021.0
pyudorandom==1.0.0
PyYAML==6.0
qdldl==0.1.5.post0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
requests-futures==1.0.0
ropwr==0.2.0
rsa==4.7.2
scipy==1.7.0
scs==3.1.0
six==1.16.0
soupsieve==2.3.0
sqlparse==0.2.0
tdigest==0.5.2.0
tenacity==8.0.0
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.0
typing_extensions==4.0.0
urllib3==1.26.0
Werkzeug==2.0.0
wrapt==1.13.0